pyproject.toml
[tool.poetry]
name = "pbg-antlr"
version = "0.1.0"
description = "Primi passi con ANTLR"
authors = ["Burlesco70 <mario.nardi.70@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
antlr4-python3-runtime = "^4.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
